+++ app/client/src/modules/ui-builder/ui/wds/WDSCustomWidget/component/styles.module.css
@@ -1,0 +1,4 @@
+.iframe {
+  width: 100%;
+  border-radius: inherit;
+}
