+++ app/client/src/modules/ui-builder/ui/wds/WDSCustomWidget/config/anvilConfig.ts
@@ -1,0 +1,9 @@
+export const anvilConfig = {
+  isLargeWidget: true,
+  widgetSize: {
+    minWidth: {
+      base: "100%",
+      "180px": "sizing-30",
+    },
+  },
+};
