+++ app/client/src/modules/ui-builder/ui/wds/WDSCustomWidget/config/index.ts
@@ -1,0 +1,6 @@
+export * from "./propertyPaneConfig";
+export { metaConfig } from "./metaConfig";
+export { anvilConfig } from "./anvilConfig";
+export { setterConfig } from "./setterConfig";
+export { defaultsConfig } from "./defaultsConfig";
+export { autocompleteConfig } from "./autocompleteConfig";
