+++ app/client/src/modules/ui-builder/ui/wds/WDSCustomWidget/config/propertyPaneConfig/index.ts
@@ -1,0 +1,2 @@
+export { propertyPaneStyleConfig } from "./styleConfig";
+export { propertyPaneContentConfig } from "./contentConfig";
