+++ app/client/src/modules/ui-builder/ui/wds/WDSCustomWidget/config/setterConfig.ts
@@ -1,0 +1,8 @@
+export const setterConfig = {
+  __setters: {
+    setVisibility: {
+      path: "isVisible",
+      type: "boolean",
+    },
+  },
+};
