+++ app/client/src/modules/ui-builder/ui/wds/WDSCustomWidget/types.ts
@@ -1,0 +1,5 @@
+import type { WidgetProps } from "widgets/BaseWidget";
+
+export interface CustomWidgetProps extends WidgetProps {
+  events: string[];
+}
